requirements.txt
gymnasium==0.29.1
stable-baselines3==2.3.2
torch==2.4.1
transformers==4.45.1
pandas==2.2.3
yfinance
pandas-datareader
python-dotenv==1.0.1
pyyaml==6.0.2
ipykernel
ipywidgets
matplotlib